volumes/bridging/config/sftp_manager/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.2.2
idna==2.6
paramiko==2.4.1
pyasn1==0.4.3
pycparser==2.18
PyNaCl==1.2.1
pysftp==0.2.9
PyYAML==3.12
six==1.11.0
